+++ src/daemon/bin/main.rs
@@ -1,11 +1,9 @@
 extern crate cartel;
-
-use cartel::daemon::api;
-use cartel::daemon::monitor;
-use cartel::daemon::signal;
-use cartel::daemon::Core;
 use std::error::Error;
-use std::sync::Arc;
+
+use cartel::daemon::bootstrap::bootstrap_shell;
+use cartel::daemon::cli::cli_app;
+use cartel::daemon::core::start_daemon;
 
 pub fn init() {
     env_logger::init()
@@ -13,10 +11,14 @@ pub fn init() {
 
 fn main() -> Result<(), Box<dyn Error>> {
     init();
-    let monitor = monitor::MonitorState::new();
-    let monitor_handle = monitor::spawn_runtime(Arc::new(monitor));
-    let core = Arc::new(Core::new(monitor_handle));
-    signal::setup_signal_handlers(Arc::clone(&core))?;
-    api::engine::start(&core);
+    let config = cli_app()?;
+    #[cfg(unix)]
+    {
+        // Re-run the daemon in a shell
+        if let Some(shell_path) = config.shell {
+            bootstrap_shell(shell_path);
+        }
+    }
+    start_daemon()?;
     Ok(())
 }